requirements.txt
Flask==3.0.0
Werkzeug==3.0.1

tensorflow==2.13.0
numpy==1.24.3
opencv-python==4.9.0.80
Pillow==10.2.0

gunicorn==21.2.0